refactor tree

+++ src/js/components/FilterTree.tsx
@@ -1,10 +1,9 @@
-import {isEqual} from "lodash"
+import {isEqual, reduce, initial, tail} from "lodash"
 import {useDispatch, useSelector} from "react-redux"
 import React from "react"
 import classNames from "classnames"
 
-import {Node} from "../models/Node"
-import {createInvestigationTree} from "./FilterTree/helpers"
+import {createInvestigationTree, InvestigationNode} from "./FilterTree/helpers"
 import {globalDispatch} from "../state/GlobalContext"
 import {submitSearch} from "../flows/submitSearch/mod"
 import BookIcon from "../icons/BookSvgIcon"
@@ -16,22 +15,79 @@ import Search from "../state/Search"
 import usePopupMenu from "./hooks/usePopupMenu"
 import {remote} from "electron"
 import Last from "../state/Last"
+import {SearchRecord} from "../types"
 
-type Props = {node: any; i: number; connId: string; spaceId: string}
+const getPins = (node?: InvestigationNode): string[] => {
+  const result = reduce(
+    node?.getPath(),
+    (res, n) => {
+      res.push(n.model.filter)
+      return res
+    },
+    []
+  )
+
+  // don't include root, or final path element as pin
+  return tail(initial(result))
+}
+
+const nodeIsPin = (node: InvestigationNode) => {
+  return node.hasChildren()
+}
+
+const nodeIsActive = (
+  prevPins: string[],
+  prevProgram: string,
+  node?: InvestigationNode
+) => {
+  return (
+    node &&
+    node.model.filter === prevProgram &&
+    isEqual(prevPins, getPins(node))
+  )
+}
+
+const reconstructSearch = (node: InvestigationNode): SearchRecord => {
+  return {
+    ...node.model.finding.search,
+    program: node.model.filter,
+    pins: getPins(node)
+  }
+}
+
+type Props = {
+  node: InvestigationNode
+  i: number
+  connId: string
+  spaceId: string
+}
 
 function NodeRow({node, i, connId, spaceId}: Props) {
   const dispatch = useDispatch()
   const last = useSelector(Last.getSearch)
-  const pinnedFilters = last?.pins || []
-  const previous = last?.program || ""
+  const prevPins = last?.pins || []
+  const prevProgram = last?.program || ""
   const menu = usePopupMenu([
     {
       label: "Delete",
       click: () => {
-        const multiTs = node.mapChildren((node) => node.data.finding.ts)
-        globalDispatch(
-          Investigation.deleteFindingByTs(connId, spaceId, multiTs)
-        )
+        remote.dialog
+          .showMessageBox({
+            type: "warning",
+            title: "Delete History Entry",
+            message: `Are you sure you want to remove this entry and it's underlying query?`,
+            buttons: ["OK", "Cancel"]
+          })
+          .then(({response}) => {
+            if (response === 0) {
+              const multiTs = node
+                .all(() => true)
+                .map((node) => node.model.finding.ts)
+              globalDispatch(
+                Investigation.deleteFindingByTs(connId, spaceId, multiTs)
+              )
+            }
+          })
       }
     },
     {type: "separator"},
@@ -55,13 +111,13 @@ function NodeRow({node, i, connId, spaceId}: Props) {
   ])
 
   function onNodeClick() {
-    dispatch(Search.restore(node.data.finding.search))
+    dispatch(Search.restore(reconstructSearch(node)))
     dispatch(submitSearch({history: true, investigation: false}))
   }
 
   const className = classNames("filter-tree-node", {
-    pinned: nodeIsPinned(pinnedFilters, node),
-    active: nodeIsActive(pinnedFilters, previous, node)
+    pinned: nodeIsPin(node),
+    active: nodeIsActive(prevPins, prevProgram, node)
   })
 
   return (
@@ -71,7 +127,7 @@ function NodeRow({node, i, connId, spaceId}: Props) {
         onClick={onNodeClick}
         onContextMenu={() => menu.open()}
       >
-        <FilterNode filter={node.data.filter} />
+        <FilterNode filter={node.model.filter} />
       </div>
       <div className="filter-tree-children">
         {node.children.map((node, i) => (
@@ -96,7 +152,7 @@ export default function FilterTree() {
   )
   const tree = createInvestigationTree(investigation)
 
-  if (tree.getRoot().children.length === 0)
+  if (tree.children.length === 0)
     return (
       <EmptySection
         icon={<BookIcon />}
@@ -106,7 +162,7 @@ export default function FilterTree() {
 
   return (
     <div className="filter-tree">
-      {tree.getRoot().children.map((node, i) => (
+      {tree.children.map((node, i) => (
         <NodeRow
           connId={currentConnId}
           spaceId={currentSpaceId}
@@ -116,46 +172,5 @@ export default function FilterTree() {
         />
       ))}
     </div>
-  )
-}
-
-export function getPinnedFilters(node: Node | null | undefined) {
-  const pinnedFilters = []
-
-  node = node && node.parent
-  while (node) {
-    if (node.isRoot()) break
-    pinnedFilters.unshift(node.data.filter)
-    node = node.parent
-  }
-
-  return pinnedFilters
-}
-
-export function nodeIsPinned(
-  pinnedFilters: string[],
-  node: Node | null | undefined
-) {
-  while (node && !node.isRoot()) {
-    const index = node.parentCount() - 1
-    const pinned = pinnedFilters[index]
-
-    if (!isEqual(node.data.filter, pinned)) return false
-
-    node = node.parent
-  }
-
-  return true
-}
-
-function nodeIsActive(
-  pinnedFilters: string[],
-  previous: string,
-  node: Node | null | undefined
-) {
-  return (
-    node &&
-    node.data.filter === previous &&
-    isEqual(pinnedFilters, getPinnedFilters(node))
   )
 }
